+++ test.ts
@@ -8,6 +8,7 @@ import { runner as notificationRunner } from "./tests/integration/notification.i
 import { runner as administrationRunner } from "./tests/integration/administration.integration";
 import { runner as oauthRunner } from "./tests/integration/oauth.integration.ts";
 import { runner as mfaRunner } from "./tests/integration/mfa.integration.ts";
+import { runner as oidcRunner } from "./tests/integration/oidc.integration";
 import { runner as platformRunner } from "./tests/integration/platform.integration";
 import { Logger } from "./tests/runner/logger";
 
@@ -36,6 +37,7 @@ async function runAllTests() {
     await administrationRunner.run(),
     await oauthRunner.run(),
     await mfaRunner.run(),
+    await oidcRunner.run(),
     await platformRunner.run(),
   ];